Create CustomImagePicker components and use it in Creation Screen
